Improve tree-sitter functionality
This changes the configuration for the tree-sitter version of todo
to move it closer to parity with the old version. In addition, it
adds a javascript file that inserts this grammar into comments
in all languages, removing the need for changes to each existing
language package (some of which may not even be maintained).